tg_bot/keyboards.py
from telegram import InlineKeyboardButton, InlineKeyboardMarkup

def main_menu_kb() -> InlineKeyboardMarkup:
    return InlineKeyboardMarkup([
        [InlineKeyboardButton('📝 Создать тикет', callback_data='menu_ticket')],
    ])
