requirements.txt
anyio==4.11.0
APScheduler==3.10.4
certifi==2025.10.5
charset-normalizer==3.4.4
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.11
PyMySQL==1.1.2
python-dotenv==1.2.1
python-telegram-bot==20.3
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wheel==0.45.1
